requirements.txt
discord.py==2.7.1
aiohttp==3.13.3
python-dotenv==1.0.0
